Make the cookie verifier stable across multiple readdirplus calls
At least one NFS client seems to expect the nfs verifier cookie
to be stable accross more than one call as the current implemenentation
does.

There are two options, the default one being using the stat
call to the dir to get the modification time-stamp and use
that as a verifier.

The advantage is that doesn't need to calculate any checksum,
the downside is, that it breaks memfs, which returns
always now() as the timestamp.