Fixed bug causing SETATTR to always fail when guard is supplied
... due to guard's ctime not matching relevant entry's Ctimespec field.

The reason why the guard's ctime will never match is due to function `ToNFSTime`
converting `time.Time.UnixNano()` to `uint32` *before* removing seconds part from
the value via a modulo operation.

Since the ctime value in the guard should be coming from a LOOKUP response (or
perhaps from some other response that contains entry stats, e.g. GETATTR or
READDIRPLUS), and the ctime field in all those responses has been "formatted"
via `ToNFSTime`, there will be no way for a remote call to SETATTR to succeed
when guard is being supplied.

This commit fixes the bug in `ToNFSTime` but also changes the guard check in the
SETATTR implementation to use the same struct as the one that is being sent in
responses that contain entry stats.  This change to the SETATTR implementation
is technically unnecessary but (IMO) it should be cleaner and safer.